[Refactor]#2 Repository 클래스들을 data.repository 패키지로 이동
